refactor: Rename namespace Serializer to Serializers
