+++ src/CoreHook.BinaryInjection/Loader/HostFunctionArguments.cs
@@ -1,5 +1,5 @@
 ﻿using System.IO;
-using CoreHook.BinaryInjection.Loader.Serializer;
+using CoreHook.BinaryInjection.Loader.Serializers;
 
 namespace CoreHook.BinaryInjection.Loader
 {
