+++ src/CoreHook.BinaryInjection/Loader/IRemoteFunctionCall.cs
@@ -1,4 +1,4 @@
-﻿using CoreHook.BinaryInjection.Loader.Serializer;
+﻿using CoreHook.BinaryInjection.Loader.Serializers;
 
 namespace CoreHook.BinaryInjection.Loader
 {
