+++ src/CoreHook.BinaryInjection/Loader/RemoteFunctionArguments.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.IO;
-using CoreHook.BinaryInjection.Loader.Serializer;
+using CoreHook.BinaryInjection.Loader.Serializers;
 
 namespace CoreHook.BinaryInjection.Loader
 {
